feat : PlaceService로 통합 및 장소명 조회 추가 (#166)
